naming + indirection fixes, added error if deleting nonexistent task

+++ app/exec/build-tasks-delete.ts
@@ -3,7 +3,7 @@
 import agentifm = require('vso-node-api/interfaces/TaskAgentInterfaces');
 import cmdm = require('../lib/tfcommand');
 import cm = require('../lib/common');
-import parameternames = require('../lib/parameternames')
+import params = require('../lib/parameternames')
 import Q = require('q');
 
 export function describe(): string {
@@ -11,7 +11,7 @@ export function describe(): string {
 }
 
 export function getArguments(): string {
-    return cmdm.formatArgumentsHint([parameternames.GENERIC_ID], []);
+    return cmdm.formatArgumentsHint([params.GENERIC_ID], []);
 }
 
 export function getCommand(): cmdm.TfCommand {
@@ -26,22 +26,28 @@ export var hideBanner: boolean = false;
 
 export class BuildTaskDelete extends cmdm.TfCommand {
     public exec(args: string[], options: cm.IOptions): any {
-        var taskId = args[0] || options[parameternames.GENERIC_ID];
-        this.checkRequiredParameter(taskId, parameternames.GENERIC_ID, parameternames.TASK_ID);
+        var taskId = args[0] || options[params.GENERIC_ID];
+        this.checkRequiredParameter(taskId, params.GENERIC_ID, params.TASK_ID);
         var deferred = Q.defer<agentifm.TaskDefinition>();
 
         var agentapi = this.getWebApi().getTaskAgentApi(this.connection.accountUrl);
 
-        //TODO: update server api to give ability to check if task exists first
-        agentapi.deleteTaskDefinition(taskId, (err, statusCode) => {
-            if(err) {
-                err.statusCode = statusCode;
-                deferred.reject(err);
+        agentapi.getTaskContent(taskId, "", (err, statusCode, tasks) => {
+            if (tasks.length > 0) {
+                agentapi.deleteTaskDefinition(taskId, (err, statusCode) => {
+                    if (err) {
+                        err.statusCode = statusCode;
+                        deferred.reject(err);
+                    }
+                    else {
+                        deferred.resolve(<agentifm.TaskDefinition>{
+                            id: taskId
+                        });
+                    }
+                });
             }
             else {
-                deferred.resolve(<agentifm.TaskDefinition>{
-                    id: taskId
-                });
+                deferred.reject(new Error("No task found with provided ID: " + taskId));
             }
         });
         return <Q.Promise<agentifm.TaskDefinition>>deferred.promise;
